Added logging and optimized functions
I added enhanced logging, can be enabled by activating -enableLogging switch by adding it to the parameters.

I created a function to add the logs to a file in CSV format.... all calls  are now logged not just the POST to Opsgenie.

Exception handling has been added to Get-SCOMalert.

Secondary redundant Get-SCOMalert has been removed, it's purpose has been included in the first Get-SCOMalert function... no need for Get-SCOMalert to be called twice.

Previously when post payload was constructed, the toString() method was being called on some objects that are already in string format.

I noticed the script send all SCOM resolution state changes , these POST request's fail on
OG end as there is no matching Alert Action.  I have therefore created a condition to suppress the alert in Powershell if the varable matches "not present" so that it is not POSTed to OG.